refactor(Core): add utilities file
[skip ci]
