--- firmware/utils.cpp
@@ -1,4 +1,0 @@
-#include <algorithm>
-#include <iterator>
-
-#include "utils.h"
