+++ lib/core/utility.hpp
